Lift CORP checks to "performa a common navigational fetch".
Though they only apply to child browsing contexts, it seems appropriate
to do these at prefetch time as long as the prefetch buffer is scoped to
the document that requested the prefetch (otherwise, the CORP checks
would occur against the wrong policy container.

In particular, this means that it is not possible to, inside a frame
which has Cross-Origin-Embedder-Policy: require-corp, navigationally
prefetch a document (or redirect to a document) which does not haave a
compatible Cross-Origin-Resource-Policy header.

This has no effect in top-level browsing contexts, but does apply to
both partitioned (N.B., due to the network partition key, this is not
actually cross-partition even if it is cross-origin) and uncredentialed
prefetch -- if either is useful in child browsing contexts, anyway.