Implement multi-file spec pull request preview on GitHub Actions. (#187)
pr-preview doesn't support multiple files (and it's regrettably difficult to add), so use GitHub Actions to replicate the feature, similar to how the deployment action works.

The results are stored in a Google Cloud Storage bucket on my personal account to which GitHub authenticates using OpenID Connect, which is quite cool and avoids the need for a secret token. The generated specs are Brotli compressed and quite small in practice.