Allow inline-speculation-rules use with nonce, hash, and strict-dynamic
"Does a source list allow all inline behavior for type?" defines
an algorithm, and it is designed to ignore 'unsafe-inline' and
'inline-speculation-rules' when one of nonce, hash, or strict-dynamic
presents.

For 'inline-speculation-rules', this is not beneficial. So, this
patch updates the spec patch not to ignore 'inline-speculation-rules'
in the algorithm.