Remove speculation rule scores from future extensions
This idea has been superseded by the eagerness key.

The eagerness section of the explainer is also updated to better
describe it as a hint instead of implying that the page can control the
browser's heuristics.

We also update a No-Vary-Search example that was using a score key to
use an equivalent eagerness value.