Fix prefetch spec list references (#277)
In the prefetch spec, references to a "list" currently link to the
structured header definition of a list, though most usage is not related
to headers. We now specify the namespace for the intentional references to
structured headers. The rest now link to the general infra definition of
a list.