Document rules: skip links which are not being rendered or in skipped contents
It's hard for the user to interact with these links, and the author may
not even intend them to be part of the document's palpable content.

In at least Chromium (and likely other implementations), these are also
elements where the author is trying to skip work related to selector
matching etc, which makes selector_matches applying to them particularly
problematic. Better to ignore them altogether.

Fixes #246.