Set prefetch URL and service worker correctly
In cases where the URL mismatches due to No-Vary-Search, we explicitly add an item to the URL list. Fixes #350.

Then, we need to ensure the active service worker is in sync with the URL of the created Document. To do this, we just manually look up the correct service worker as part of creating the reserved client.