Set prefetch URL and service worker correctly
This implements the current consensus outlined in #388:

- Only the No-Vary-Search of the first response in the redirect chain is
used. Previously, we were using the No-Vary-Search of the last response,
which was a pretty bad mismatch with the intended semantics and perhaps
a security issue. Fixes #388.

- We treat No-Vary-Search as inserting a kind of internal redirect into
the URL list right after the first response, so that if there are no
redirects, the navigated-to URL is used instead of the original response
URL. But if there are redirects, then the final destination URL is used.
Fixes #350.

- For determining the active service worker, we carry over the one that
controlled the last response of the prefetch.

All of these decisions are explained in new notes in the spec text.

As discussed in #388, this solution is not guaranteed to work
(especially for service workers) if we allow prefetches to be used to
fulfill redirect responses. The spec currently does not allow that, but
Chromium does; that mismatch is tracked in #367. We will need to ensure
any solution toward #367 preserves the desired properties, possibly
including an explicit bail-out if we see a combination of prefetch +
service workers + redirects.